+++ lib/src/event_phase.dart
@@ -11,6 +11,8 @@ enum EventPhase {
 
   yielded,
 
+  interrupted,
+
   finished;
 
   /// Returns the string representation of the status.