+++ lib/src/internal/completer_interrupt.dart
@@ -1,0 +1,6 @@
+import 'package:meta/meta.dart';
+
+@internal
+class CompleterInterrupt{
+
+}