+++ lib/src/internal/debug_listener.dart
@@ -1,0 +1,18 @@
+import 'package:meta/meta.dart';
+
+@internal
+abstract class DebugListener {
+
+  void onScheduleNextAction();
+
+  void onNextAction();
+  
+  void onExecuteAction();
+
+  void onStop();
+
+  void onAddCompleter();
+
+  void onRemoveCompleter();
+}
+