+++ lib/src/internal/event_action.dart
@@ -4,6 +4,8 @@ import 'package:meta/meta.dart';
 import 'package:simdart/src/event.dart';
 import 'package:simdart/src/event_phase.dart';
 import 'package:simdart/src/internal/completer_action.dart';
+import 'package:simdart/src/internal/completer_interrupt.dart';
+import 'package:simdart/src/internal/stop_action.dart';
 import 'package:simdart/src/internal/time_action.dart';
 import 'package:simdart/src/interval.dart';
 import 'package:simdart/src/resources.dart';
@@ -48,50 +50,60 @@ class EventAction extends TimeAction implements SimContext {
 
   @override
   void execute() {
+    print('execute: $eventName');
     if (_eventCompleter != null) {
       throw StateError('This event is yielding');
     }
 
-    sim.observer?.onEvent(
+    sim.listener?.onEvent(
         name: eventName,
         time: sim.now,
         phase: EventPhase.called,
         executionHash: hashCode);
 
-    _runEvent().then((_) {
-      if (_eventCompleter != null) {
-        SimDartHelper.error(
-            sim: sim,
-            msg:
-                "Next event is being scheduled, but the current one is still paused waiting for continuation. Did you forget to use 'await'?");
-        return;
+    event(this).then((_) {
+      print('then $eventName ${sim.runState.name}');
+        if (_eventCompleter != null && sim.runState==RunState.running) {
+          //TODO method or throw?
+          SimDartHelper.error(
+              sim: sim,
+              error:
+              StateError(
+                  "Next event is being scheduled, but the current one is still paused waiting for continuation. Did you forget to use 'await'?"));
+          return;
+        }
+        sim.listener?.onEvent(
+            name: eventName,
+            time: sim.now,
+            phase: EventPhase.finished,
+            executionHash: hashCode);
+        SimDartHelper.scheduleNextAction(sim: sim);
+    }).catchError((error) {
+      //TODO rever
+      if(error is CompleterInterrupt) {
+        sim.listener?.onEvent(
+            name: eventName,
+            time: sim.now,
+            phase: EventPhase.interrupted,
+            executionHash: hashCode);
+      } else  {
+        print('ZZ');
+        SimDartHelper.error(sim: sim, error: error);
       }
-      sim.observer?.onEvent(
-          name: eventName,
-          time: sim.now,
-          phase: EventPhase.finished,
-          executionHash: hashCode);
-
-      SimDartHelper.scheduleNextAction(sim: sim);
-    }).catchError((e) {
-      // Sim already marked to finish. Let the last event finalize.
     });
-  }
-
-  Future<void> _runEvent() async {
-    await event(this);
   }
 
   @override
   Future<void> wait(int delay) async {
     if (_eventCompleter != null) {
+      //TODO method or throw?
       SimDartHelper.error(
           sim: sim,
-          msg: "The event is already waiting. Did you forget to use 'await'?");
+          error: StateError("The event is already waiting. Did you forget to use 'await'?"));
       return;
     }
 
-    sim.observer?.onEvent(
+    sim.listener?.onEvent(
         name: eventName,
         time: sim.now,
         phase: EventPhase.yielded,
@@ -106,9 +118,9 @@ class EventAction extends TimeAction implements SimContext {
             start: sim.now + delay,
             complete: _eventCompleter!.complete,
             order: order));
-    SimDartHelper.scheduleNextAction(sim: sim);
 
-    await _eventCompleter!.future;
+    SimDartHelper.scheduleNextAction(sim: sim);
+      await _eventCompleter!.future;
   }
 
   @override
@@ -134,6 +146,11 @@ class EventAction extends TimeAction implements SimContext {
   }
 
   @override
+  void stop(){
+    SimDartHelper.addAction(sim: sim, action: StopAction(start: sim.now, sim:sim));
+  }
+
+  @override
   SimCounter counter(String name) {
     return sim.counter(name);
   }
@@ -142,15 +159,12 @@ class EventAction extends TimeAction implements SimContext {
   SimNum num(String name) {
     return sim.num(name);
   }
-
-  @override
-  void dispose() {
-    _eventCompleter?.complete();
-  }
 }
 
 class EventCompleter {
-  EventCompleter({required this.event});
+  EventCompleter({required this.event}){
+    SimDartHelper.addCompleter(sim: event.sim, completer: _completer);
+  }
 
   final Completer<void> _completer = Completer();
 
@@ -159,12 +173,13 @@ class EventCompleter {
   Future<void> get future => _completer.future;
 
   void complete() {
-    event.sim.observer?.onEvent(
+    event.sim.listener?.onEvent(
         name: event.eventName,
         time: event.sim.now,
         phase: EventPhase.resumed,
         executionHash: hashCode);
     _completer.complete();
     event._eventCompleter = null;
+    SimDartHelper.removeCompleter(sim: event.sim, completer: _completer);
   }
 }