+++ lib/src/internal/time_action.dart
@@ -14,6 +14,4 @@ abstract class TimeAction {
   final int order;
 
   void execute();
-
-  void dispose();
 }