+++ test/resource_test.dart
@@ -8,7 +8,7 @@ void main() {
   TestHelper helper = TestHelper();
 
   setUp(() {
-    sim = SimDart(observer: helper);
+    sim = SimDart(listener: helper);
   });
 
   group('Resource', () {
@@ -39,7 +39,7 @@ void main() {
 
       await sim.run();
 
-      helper.test([
+      helper.testEvents([
         '[0][A][called]',
         '[0][A][yielded]',
         '[0][B][called]',
@@ -73,7 +73,7 @@ void main() {
 
       await sim.run();
 
-      helper.test([
+      helper.testEvents([
         '[0][A][called]',
         '[0][A][yielded]',
         '[0][B][called]',
@@ -110,7 +110,7 @@ void main() {
 
       await sim.run();
 
-      helper.test([
+      helper.testEvents([
         '[0][A][called]',
         '[0][A][yielded]',
         '[1][B][called]',
@@ -143,12 +143,7 @@ void main() {
           sim.process(event: (context) async {});
           await sim.run();
         },
-        throwsA(
-          predicate((e) =>
-              e is StateError &&
-              e.message.contains(
-                  "This event should be waiting for the resource to be released. Did you forget to use 'await'?")),
-        ),
+          throwsA(isA<StateError>().having((e) => e.message, 'message', equals("This event should be waiting. Did you forget to use 'await'?")))
       );
     });
   });