+++ test/time_loop_test.dart
@@ -19,9 +19,6 @@ class TestAction extends TimeAction {
     names.add(name);
     SimDartHelper.scheduleNextAction(sim: sim);
   }
-
-  @override
-  void dispose() {}
 }
 
 void main() {