+++ test/wait_test.dart
@@ -10,7 +10,7 @@ void main() {
   TestHelper helper = TestHelper();
 
   setUp(() {
-    sim = SimDart(observer: helper);
+    sim = SimDart(listener: helper);
   });
 
   group('Wait', () {
@@ -38,7 +38,7 @@ void main() {
 
       await sim.run();
 
-      helper.test([
+      helper.testEvents([
         '[0][a][called]',
         '[0][a][yielded]',
         '[10][a][resumed]',
@@ -54,7 +54,7 @@ void main() {
       sim.process(event: emptyEvent, start: 5, name: 'b');
 
       await sim.run();
-      helper.test([
+      helper.testEvents([
         '[0][a][called]',
         '[0][a][yielded]',
         '[5][b][called]',
@@ -74,7 +74,7 @@ void main() {
       sim.process(event: emptyEvent, delay: 5, name: 'b');
 
       await sim.run();
-      helper.test([
+      helper.testEvents([
         '[0][a][called]',
         '[0][a][yielded]',
         '[5][b][called]',
@@ -98,12 +98,7 @@ void main() {
 
           await sim.run();
         },
-        throwsA(
-          predicate((e) =>
-              e is StateError &&
-              e.message.contains(
-                  "Next event is being scheduled, but the current one is still paused waiting for continuation. Did you forget to use 'await'?")),
-        ),
+          throwsA(isA<StateError>().having((e) => e.message, 'message', equals("Next event is being scheduled, but the current one is still paused waiting for continuation. Did you forget to use 'await'?")))
       );
     });
 
@@ -117,13 +112,9 @@ void main() {
               },
               name: 'a');
           await sim.run();
+          print('depois');
         },
-        throwsA(
-          predicate((e) =>
-              e is StateError &&
-              e.message.contains(
-                  "The event is already waiting. Did you forget to use 'await'?")),
-        ),
+        throwsA(isA<StateError>().having((e) => e.message, 'message', equals("The event is already waiting. Did you forget to use 'await'?")))
       );
     });
   });